+++ src/IRProtocol.hpp
@@ -81,8 +81,8 @@ const char string_FAST[] PROGMEM = "FAST";
 const char *const ProtocolNames[]
 PROGMEM = { string_Unknown, string_PulseWidth, string_PulseDistance, string_Apple, string_Denon, string_JVC, string_LG, string_LG2,
         string_NEC, string_NEC2, string_Onkyo, string_Panasonic, string_Kaseikyo, string_Kaseikyo_Denon, string_Kaseikyo_Sharp,
-        string_Kaseikyo_JVC, string_Kaseikyo_Mitsubishi, string_RC5, string_RC6, string_RC6A, string_Samsung, string_SamsungLG, string_Samsung48,
-        string_Sharp, string_Sony
+        string_Kaseikyo_JVC, string_Kaseikyo_Mitsubishi, string_RC5, string_RC6, string_RC6A, string_Samsung, string_SamsungLG,
+        string_Samsung48, string_Sharp, string_Sony
 #if !defined(EXCLUDE_EXOTIC_PROTOCOLS)
         , string_BangOlufsen, string_BoseWave, string_Lego, string_MagiQuest, string_Whynter, string_FAST
 #endif
@@ -160,7 +160,7 @@ namespace PrintULL {
  * @param aPrintRepeatGap   If true also print the gap before repeats.
  *
  */
-void printIRResultShort(Print *aSerial, IRData *aIRDataPtr, bool aPrintRepeatGap) {
+void printIRResultShort(Print *aSerial, IRData *aIRDataPtr) {
     if (aIRDataPtr->flags & IRDATA_FLAGS_WAS_OVERFLOW) {
         aSerial->println(F("Overflow"));
         return;
@@ -210,22 +210,6 @@ void printIRResultShort(Print *aSerial, IRData *aIRDataPtr, bool aPrintRepeatGap
 #if defined(DECODE_DISTANCE_WIDTH)
         }
 #endif
-        if (aIRDataPtr->flags & (IRDATA_FLAGS_IS_AUTO_REPEAT | IRDATA_FLAGS_IS_REPEAT)) {
-            aSerial->print(' ');
-            if (aIRDataPtr->flags & IRDATA_FLAGS_IS_AUTO_REPEAT) {
-                aSerial->print(F("Auto-"));
-            }
-            aSerial->print(F("Repeat"));
-#if !defined(DISABLE_CODE_FOR_RECEIVER)
-            if (aPrintRepeatGap) {
-                aSerial->print(F(" gap="));
-                aSerial->print((uint32_t) aIRDataPtr->initialGapTicks * MICROS_PER_TICK);
-                aSerial->print(F("us"));
-            }
-#else
-            (void)aPrintRepeatGap;
-#endif
-        }
 
         /*
          * Print raw data
@@ -245,14 +229,37 @@ void printIRResultShort(Print *aSerial, IRData *aIRDataPtr, bool aPrintRepeatGap
             aSerial->print(F(" bits"));
 
             if (aIRDataPtr->flags & IRDATA_FLAGS_IS_MSB_FIRST) {
-                aSerial->println(F(" MSB first"));
+                aSerial->print(F(" MSB first"));
             } else {
-                aSerial->println(F(" LSB first"));
+                aSerial->print(F(" LSB first"));
             }
-
-        } else {
-            aSerial->println();
         }
+
+        /*
+         * Print gap and duration, in order to be able to compute the repeat period of the protocol by adding the next gap time
+         */
+        if (aIRDataPtr->flags & (IRDATA_FLAGS_IS_AUTO_REPEAT | IRDATA_FLAGS_IS_REPEAT)) {
+            aSerial->print(' ');
+            if (aIRDataPtr->flags & IRDATA_FLAGS_IS_AUTO_REPEAT) {
+                aSerial->print(F("Auto-"));
+            }
+            aSerial->print(F("Repeat"));
+        }
+#if !defined(DISABLE_CODE_FOR_RECEIVER)
+        aSerial->print(F(" Gap="));
+        aSerial->print((uint32_t) aIRDataPtr->initialGapTicks * MICROS_PER_TICK);
+        aSerial->print(F("us"));
+
+        uint16_t tSumOfDurationTicks = 0;
+        for (IRRawlenType i = 1; i < aIRDataPtr->rawlen; i++) {
+            tSumOfDurationTicks += aIRDataPtr->rawDataPtr->rawbuf[i];
+        }
+        aSerial->print(F(" Duration="));
+        aSerial->print((uint32_t) tSumOfDurationTicks * MICROS_PER_TICK, DEC);
+        aSerial->println(F("us"));
+#else
+        aSerial->println();
+#endif
     }
 }