+++ changelog.md
@@ -5,6 +5,8 @@ See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-I
 # 4.4.2
 - Support for SAMD51 timer3 if timer 5 is not available (Adafruit ItsyBitsy M4).
 - attachInterrupt() on SAMD has a different semantic :-(. See: https://www.arduino.cc/reference/tr/language/functions/external-interrupts/attachinterrupt/.
+- Fixed overflow handling.
+- Improved repeat detection for DistanceWidthProtocol.
 
 # 4.4.1
 - Support for ESP core 3.x by akellai.